fix: always use forward slashes in secret names
Currently, running `gopass list --flat` on Windows on a store containing a
single secret

    a/long/secret/name/with/slashes

produces this output:

```
C:\> gopass list --flat
a\long\secret\name\with/slashes
```

This is problematic when the output of that call is used programmatically
later, e.g. to feed a `gopass show` call.

With this patch, backslashes are replaced with forward slashes when listing
secret names. This should have no impact on platforms that use `/` as a path
separator.

Signed-off-by: Florian Esser <fe@wetransform.to>